requirement.txt
pyautogui
Pillow
pygetwindow
psutil
pywin32
opencv-python
mouse
pytesseract